Begin protocol refactor
